improve irit io tests.
